requirements.txt
python-dotenv
tweepy
litellm
pyyaml
markdown

